refactor(backend): colocate imports domain into feature package
Moves the vanilla-import controller and UUID hasher utility out of the
generic controller/utils layer packages into
net.onelitefeather.blackhole.backend.imports, alongside the
already-colocated VanillaImportService and its DTOs. Pure physical
move, no behavior change.
